Move test functions to `test_stand`
There they can be re-used by other driver test suites.
